Made Stockchart a seperate component, created corresponding css styling and changed to waterfall chart with different options and sample data.
