+++ twitterlens/src/App.css
@@ -14,8 +14,8 @@
 }
 
 .App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+  background-color: #2f3238;
+  min-height: 20vh;
   display: flex;
   flex-direction: column;
   align-items: center;
