+++ twitterlens/src/components/Stockchart.css
@@ -1,0 +1,14 @@
+.Stockchart {
+    text-align: center;
+}
+
+.Stock-header {
+    background-color: #282c34;
+    min-height: 80vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: normal;
+    font-size: calc(10px + 2vmin);
+    color: white;
+  }
